requirements.txt

Django==3.2
django-appconf==1.0.4
django-autoslug==1.9.8
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
Jinja2==2.11.3
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3